@@ -1,5 +1,5 @@
 name: qiniu_sdk_base
-version: 0.4.0
+version: 0.5.0
 homepage: https://github.com/qiniu/dart-sdk
 description: The sdk basic of Qiniu products
 
@@ -7,17 +7,17 @@ false_secrets:
   - /test/auth/auth_test.dart
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.15.0 <4.0.0'
 
 dependencies:
-  dio: ^4.0.0
+  dio: ^5.0.0
   crypto: ^3.0.0
   meta: ^1.3.0
   uuid: ^3.0.5
   path: ^1.8.0
 
 dev_dependencies:
-  pedantic: ^1.11.0
-  test: ^1.16.8
-  dotenv: 3.0.0-nullsafety.0
-  coverage: "^1.0.2"
+  test:
+  lints:
+  dotenv: ^4.1.0
+  coverage: "^1.6.3"
