@@ -1,20 +1,20 @@
 name: qiniu_flutter_sdk
 description: Qiniu Flutter sdk
-version: 0.4.0
+version: 0.5.0
 homepage: https://github.com/qiniu/dart-sdk/tree/master/flutter
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.15.0 <4.0.0"
+  flutter: ">=2.8.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  dio: ^4.0.0
-  qiniu_sdk_base: ^0.4.0
+  dio: ^5.0.0
+  qiniu_sdk_base: ^0.5.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  flutter_lints:
   flutter_test:
     sdk: flutter
 
