@@ -1,9 +1,6 @@
 import 'package:dio/dio.dart';
 
-import 'package:qiniu_sdk_base/src/storage/error/error.dart';
-
-import 'http_adaptor/native_adaptor.dart'
-    if (dart.library.html) 'http_adaptor/browser_adaptor.dart';
+import '../../storage/error/error.dart';
 
 part 'cache.dart';
 part 'host.dart';
@@ -26,5 +23,5 @@ class Config {
     this.retryLimit = 3,
   })  : hostProvider = hostProvider ?? DefaultHostProvider(),
         cacheProvider = cacheProvider ?? DefaultCacheProvider(),
-        httpClientAdapter = httpClientAdapter ?? createHttpAdaptor();
+        httpClientAdapter = httpClientAdapter ?? HttpClientAdapter();
 }
