@@ -30,6 +30,8 @@ class PutOptions {
     this.maxPartsRequestNumber = 5,
     this.customVars,
     this.controller,
-  }) : assert(partSize >= 1 && partSize <= 1024,
-            'partSize must be greater than or equal to 1 and less than or equal to 1024');
+  }) : assert(
+          partSize >= 1 && partSize <= 1024,
+          'partSize must be greater than or equal to 1 and less than or equal to 1024',
+        );
 }
