@@ -1,6 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:meta/meta.dart';
-import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
+
+import '../../../qiniu_sdk_base.dart';
 
 part 'request_task_controller.dart';
 part 'request_task_manager.dart';
@@ -52,15 +53,27 @@ abstract class RequestTask<T> extends Task<T> {
     controller?.notifyProgressListeners(preStartTakePercentOfTotal);
     retryLimit = config.retryLimit;
     client.httpClientAdapter = config.httpClientAdapter;
-    client.interceptors.add(InterceptorsWrapper(onRequest: (options, handler) {
-      controller?.notifyStatusListeners(StorageStatus.Request);
-      options
-        ..cancelToken = controller?.cancelToken
-        ..onSendProgress = (sent, total) => onSendProgress(sent / total);
-      options.headers['User-Agent'] = _getUserAgent();
+    client.interceptors.add(
+      InterceptorsWrapper(
+        onRequest: (options, handler) {
+          controller?.notifyStatusListeners(StorageStatus.Request);
+          options
+            ..cancelToken = controller?.cancelToken
+            ..onSendProgress = (sent, total) => onSendProgress(sent / total);
+          options.headers['User-Agent'] = _getUserAgent();
 
-      handler.next(options);
-    }));
+          if (options.contentType == null) {
+            if (options.data is Stream) {
+              options.contentType = 'application/octet-stream';
+            } else {
+              options.contentType = 'application/json';
+            }
+          }
+
+          handler.next(options);
+        },
+      ),
+    );
 
     super.preStart();
   }
@@ -152,7 +165,7 @@ abstract class RequestTask<T> extends Task<T> {
     if (!_canConnectToHost(error) || _isHostUnavailable(error)) {
       return true;
     }
-    if (error.type == DioErrorType.response) {
+    if (error.type == DioErrorType.badResponse) {
       if (error.response?.statusCode == 612) {
         return true;
       }
@@ -163,7 +176,7 @@ abstract class RequestTask<T> extends Task<T> {
   // host 是否可以连接上
   bool _canConnectToHost(Object error) {
     if (error is DioError) {
-      if (error.type == DioErrorType.response) {
+      if (error.type == DioErrorType.badResponse) {
         final statusCode = error.response?.statusCode;
         if (statusCode is int && statusCode > 99) {
           return true;
@@ -181,7 +194,7 @@ abstract class RequestTask<T> extends Task<T> {
   // host 是否不可用
   bool _isHostUnavailable(Object error) {
     if (error is DioError) {
-      if (error.type == DioErrorType.response) {
+      if (error.type == DioErrorType.badResponse) {
         final statusCode = error.response?.statusCode;
         if (statusCode == 502) {
           return true;
