@@ -1,7 +1,8 @@
 import 'dart:async';
 
 import 'package:meta/meta.dart';
-import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
+
+import '../../../qiniu_sdk_base.dart';
 
 export 'request_task.dart';
 export 'task_manager.dart';
