@@ -1,5 +1,3 @@
-// ignore: import_of_legacy_library_into_null_safe
-import 'package:dotenv/dotenv.dart';
 import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
 import 'package:test/test.dart';
 
@@ -7,19 +5,23 @@ import 'config.dart';
 
 void main() {
   configEnv();
-  test('RegionProvider should works well.', () async {
-    final hostProvider = DefaultHostProvider();
-    final tokenInfo = Auth.parseUpToken(token);
-    final putPolicy = tokenInfo.putPolicy;
+  test(
+    'RegionProvider should works well.',
+    () async {
+      final hostProvider = DefaultHostProvider();
+      final tokenInfo = Auth.parseUpToken(token);
+      final putPolicy = tokenInfo.putPolicy;
 
-    final hostInToken = await hostProvider.getUpHost(
-      accessKey: tokenInfo.accessKey,
-      bucket: putPolicy.getBucket(),
-    );
+      final hostInToken = await hostProvider.getUpHost(
+        accessKey: tokenInfo.accessKey,
+        bucket: putPolicy.getBucket(),
+      );
 
-    // 根据传入的 token 的 bucket 对应的区域，需要对应的修改这里
-    expect(hostInToken, 'https://upload-na0.qiniup.com');
-  }, skip: !isSensitiveDataDefined);
+      // 根据传入的 token 的 bucket 对应的区域，需要对应的修改这里
+      expect(hostInToken, 'https://upload-na0.qiniup.com');
+    },
+    skip: !isSensitiveDataDefined,
+  );
 
   test('DefaultCacheProvider should works well.', () async {
     final cacheProvider = DefaultCacheProvider();
@@ -35,20 +37,23 @@ void main() {
     expect(cacheProvider.value.length, 0);
   });
 
-  test('freeze mechanism should works well with DefaultHostProvider.',
-      () async {
-    final config = Config();
-    final hostA = await config.hostProvider.getUpHost(
-      accessKey: env['QINIU_DART_SDK_ACCESS_KEY']!,
-      bucket: env['QINIU_DART_SDK_TOKEN_SCOPE']!,
-    );
-    config.hostProvider.freezeHost(hostA);
-    final hostB = await config.hostProvider.getUpHost(
-      accessKey: env['QINIU_DART_SDK_ACCESS_KEY']!,
-      bucket: env['QINIU_DART_SDK_TOKEN_SCOPE']!,
-    );
+  test(
+    'freeze mechanism should works well with DefaultHostProvider.',
+    () async {
+      final config = Config();
+      final hostA = await config.hostProvider.getUpHost(
+        accessKey: env['QINIU_DART_SDK_ACCESS_KEY']!,
+        bucket: env['QINIU_DART_SDK_TOKEN_SCOPE']!,
+      );
+      config.hostProvider.freezeHost(hostA);
+      final hostB = await config.hostProvider.getUpHost(
+        accessKey: env['QINIU_DART_SDK_ACCESS_KEY']!,
+        bucket: env['QINIU_DART_SDK_TOKEN_SCOPE']!,
+      );
 
-    // getUpHost 会返回至少2个host，不用担心会少于两个
-    expect(hostA == hostB, false);
-  }, skip: !isSensitiveDataDefined);
+      // getUpHost 会返回至少2个host，不用担心会少于两个
+      expect(hostA == hostB, false);
+    },
+    skip: !isSensitiveDataDefined,
+  );
 }
