@@ -1,7 +1,6 @@
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:dio/adapter.dart';
 import 'package:dio/dio.dart';
 import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
 import 'package:test/test.dart';
@@ -57,19 +56,22 @@ class PutControllerBuilder {
   }
 }
 
-class HttpAdapterTestWith612 extends HttpClientAdapter {
+class HttpAdapterTestWith612 implements HttpClientAdapter {
   /// 记录 CompletePartsTask 被创建的次数
   /// 第一次我们拦截并返回 612，第二次不拦截
   bool completePartsTaskResponse612 = false;
-  final DefaultHttpClientAdapter _adapter = DefaultHttpClientAdapter();
+  final HttpClientAdapter _adapter = HttpClientAdapter();
   @override
   void close({bool force = false}) {
     _adapter.close(force: force);
   }
 
   @override
-  Future<ResponseBody> fetch(RequestOptions options,
-      Stream<Uint8List>? requestStream, Future? cancelFuture) async {
+  Future<ResponseBody> fetch(
+    RequestOptions options,
+    Stream<Uint8List>? requestStream,
+    Future? cancelFuture,
+  ) async {
     /// 如果是 CompletePartsTask 发出去的请求，则返回 612
     if (options.path.contains('uploads/') &&
         options.method == 'POST' &&
