@@ -1,2 +1,2 @@
-final String? builtinToken =
+const builtinToken =
     'lVgtk5xr03Oz_uvkzDtQ8LtpiEUWx5tGEDUZVg1y:KBGQR5DaE7J03MYSrlF9kxtE8EY=:eyJyZXR1cm5Cb2R5Ijoie1wia2V5XCI6ICQoa2V5KX0iLCJzY29wZSI6InFpbml1LWRhcnQtc2RrIiwiZGVhZGxpbmUiOjE2MTg4NjIxMDZ9';
