@@ -1,14 +1,15 @@
 import 'package:flutter/material.dart';
 
 class Progress extends StatelessWidget {
+  const Progress(this.value, {Key? key}) : super(key: key);
+
   final double value;
-  Progress(this.value);
 
   Widget get progress {
     return Container(
       width: 200,
       height: 200,
-      padding: EdgeInsets.all(20),
+      padding: const EdgeInsets.all(20),
       child: CircularProgressIndicator(
         value: value,
         strokeWidth: 14,
@@ -20,10 +21,10 @@ class Progress extends StatelessWidget {
   Widget get progressText {
     final integer = (value * 100).toInt();
     final decimal = (((value * 100) - integer) * 100).toInt();
-    final unit = '%';
+    const unit = '%';
 
     return DefaultTextStyle(
-        style: TextStyle(
+        style: const TextStyle(
           color: Colors.black87,
           fontWeight: FontWeight.bold,
         ),
@@ -32,12 +33,12 @@ class Progress extends StatelessWidget {
           children: [
             Text(
               integer.toString(),
-              style: TextStyle(fontSize: 50),
+              style: const TextStyle(fontSize: 50),
             ),
             Column(
               children: [
-                Text(unit.toString(), style: TextStyle(fontSize: 14)),
-                Text('.$decimal', style: TextStyle(fontSize: 18)),
+                Text(unit.toString(), style: const TextStyle(fontSize: 14)),
+                Text('.$decimal', style: const TextStyle(fontSize: 18)),
               ],
             )
           ],
