@@ -14,16 +14,18 @@ export 'package:qiniu_sdk_base/qiniu_sdk_base.dart'
         QiniuError,
         StorageError,
         StorageErrorType,
-        PutByPartOptions,
-        StorageStatus,
-        PutBySingleOptions;
+        StorageStatus;
 
 export './controller.dart';
 
+/// Storage
 class Storage {
-  final base.Storage _baseStorage;
+  /// Storage
   Storage({base.Config? config}) : _baseStorage = base.Storage(config: config);
 
+  final base.Storage _baseStorage;
+
+  /// putFile
   Future<base.PutResponse> putFile(
     File file,
     String token, {
@@ -32,6 +34,7 @@ class Storage {
     return _baseStorage.putFile(file, token, options: options);
   }
 
+  /// putBytes
   Future<base.PutResponse> putBytes(
     Uint8List bytes,
     String token, {
